Maintain database connections for Gunicorn
When ha_server is run under Gunicorn the database is not
initialized in main.  Therefore, database connections need
to be setup for each request.  This requires further adjustments
on how testing is performed.  A configuration is set preventing
database connections from getting dropped after each test, so that
the database connents can be examined.